Add MeshDG mixin
